refactor : 지하철 노선에 구간 추가하는 로직에 대한 테스트를 따로 파일로 분리
