+++ TODO.md
@@ -64,7 +64,7 @@
 
 ### 기타
 - [ ] 구간 등록에 대한 ATDD 테스트 코드 추가하기
-- [ ] RestAssured에서 QueryParams, PathParams 활용해서 리팩토링하기
+- [x] RestAssured에서 QueryParams, PathParams 활용해서 리팩토링하기
 - [x] 예외 처리를 커스텀 예외 활용하도록 바꾸기
 - [ ] Service Layer에 있는 비즈니스 로직을 Domain으로 옮기기
 - [ ] ATDD에서는 @Transactional이 적용이 안 되는 지 테스트하기 / h2 DB에서는 Truncate로는 AI가 초기화 안 되는 건 아닌지
